refactor: extract dataclasses to models module
